requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.3
django-appconf==0.6
django-compressor==1.4
django-filter==0.9.2
djangorestframework==3.0.4
djangorestframework-jwt==1.2.0
drf-nested-routers==0.9.0
gunicorn==19.2.1
httpie==0.9.1
Markdown==2.5.2
mysqlclient==1.3.4
Pillow==2.7.0
Pygments==2.0.2
PyJWT==0.4.1
requests==2.5.1
six==1.9.0
static3==0.5.1
virtualenv==12.0.7
